refactor: ChartDataCommand into two separate commands (#17425)
