+++ superset/tasks/async_queries.py
@@ -55,7 +55,7 @@ def load_chart_data_into_cache(
     job_metadata: Dict[str, Any], form_data: Dict[str, Any],
 ) -> None:
     # pylint: disable=import-outside-toplevel
-    from superset.charts.commands.data import ChartDataCommand
+    from superset.charts.data.commands import ChartDataCommand
 
     try:
         ensure_user_is_set(job_metadata.get("user_id"))
